RL.Model

Documentation

activation :: forall (dev :: (DeviceType, Nat)) (shape :: [Nat]). IsValidDevice dev => QTensor dev shape -> QTensor dev shape

expandAs :: Tensor -> Tensor -> Tensor

traceDyn :: forall (a :: (DeviceType, Nat)) (b :: DType) (c :: [Nat]). Tensor a b c -> Tensor a b c

unsafeReshape :: forall (dev :: (DeviceType, Nat)) (dtype :: DType) (shape :: [Nat]) (shape' :: [Nat]). [Int] -> Tensor dev dtype shape -> Tensor dev dtype shape'

data ConstEmbSpec (dev :: k) (shape :: [Nat])

Constructors

ConstEmbSpec

newtype ConstEmb (dev :: (DeviceType, Nat)) (shape :: [Nat])

Constructors

ConstEmb (Parameter dev QDType shape)

data SliceSpec (dev :: k)

Constructors

SliceSpec

data SliceEncoder (dev :: (DeviceType, Nat))

Constructors

SliceEncoder

Fields

_slcL1 :: !(Conv2d 1 QSliceHidden 1 1 QDType dev)

_slcL2 :: !(Conv2d QSliceHidden EmbSize FifthSize OctaveSize QDType dev)

_slcStart :: !(ConstEmb dev EmbShape)

_slcStop :: !(ConstEmb dev EmbShape)

data TransitionSpec (dev :: k)

Constructors

TransitionSpec

data TransitionEncoder (dev :: (DeviceType, Nat))

Constructors

TransitionEncoder

Fields

trL1Passing :: !(Conv2d 2 QTransHidden FifthSize OctaveSize QDType dev)

trL1Inner :: !(Conv2d 2 QTransHidden FifthSize OctaveSize QDType dev)

trL1Left :: !(Conv2d 1 QTransHidden 1 1 QDType dev)

trL1Right :: !(Conv2d 1 QTransHidden 1 1 QDType dev)

trL1Root :: !(ConstEmb dev '[QTransHidden])

trL2 :: !(Conv2d QTransHidden EmbSize FifthSize OctaveSize QDType dev)

data ActionSpec (dev :: k)

Constructors

ActionSpec

data ActionEncoder (dev :: (DeviceType, Nat))

Constructors

ActionEncoder

Fields

actTop1sl :: !(Conv2d EmbSize QActionHidden FifthSize OctaveSize QDType dev)

actTop1sm :: !(Conv2d EmbSize QActionHidden FifthSize OctaveSize QDType dev)

actTop1sr :: !(Conv2d EmbSize QActionHidden FifthSize OctaveSize QDType dev)

actTop1t1 :: !(Conv2d EmbSize QActionHidden FifthSize OctaveSize QDType dev)

actTop1t2 :: !(Conv2d EmbSize QActionHidden FifthSize OctaveSize QDType dev)

actTop2 :: !(Conv2d QActionHidden EmbSize FifthSize OctaveSize QDType dev)

actSplit :: ConstEmb dev '[EmbSize - 3]

actSpread :: ConstEmb dev '[EmbSize - 3]

actFreeze :: ConstEmb dev '[EmbSize - 3]

opTypes :: forall (dev :: (DeviceType, Nat)). KnownDevice dev => QTensor dev '[6, 3]

data StateSpec (dev :: k)

Constructors

StateSpec

data StateEncoder (dev :: (DeviceType, Nat))

Constructors

StateEncoder

Fields

stL1mid :: Conv2d EmbSize QStateHidden FifthSize OctaveSize QDType dev

stL1frozenSlc :: Conv2d EmbSize QStateHidden FifthSize OctaveSize QDType dev

stL1frozenTr :: Conv2d EmbSize QStateHidden FifthSize OctaveSize QDType dev

stL1openSlc :: Conv2d EmbSize QStateHidden FifthSize OctaveSize QDType dev

stL1openTr :: Conv2d EmbSize QStateHidden FifthSize OctaveSize QDType dev

stL2 :: Conv2d QStateHidden QStateHidden FifthSize OctaveSize QDType dev

stL3 :: Conv2d QStateHidden EmbSize FifthSize OctaveSize QDType dev

data QSpec (dev :: k)

Constructors

QSpec

data QModel (dev :: (DeviceType, Nat))

Constructors

QModel

Fields

qModelSlc :: !(SliceEncoder dev)

qModelTr :: !(TransitionEncoder dev)

qModelAct :: !(ActionEncoder dev)

qModelSt :: !(StateEncoder dev)

qModelFinal1 :: !(Conv2d EmbSize QOutHidden FifthSize OctaveSize QDType dev)

qModelNorm1 :: !(LayerNorm '[QOutHidden] QDType dev)

qModelFinal2 :: !(Linear QOutHidden 1 QDType dev)

qModelValue1 :: !(Linear EmbSize QOutHidden QDType dev)

qModelValueNorm :: !(LayerNorm '[QOutHidden] QDType dev)

qModelValue2 :: !(Linear QOutHidden 1 QDType dev)

forwardQModel :: forall (dev :: (DeviceType, Nat)). IsValidDevice dev => QModel dev -> QEncoding dev ('[] :: [Nat]) -> QTensor dev '[1]

forwardQModelBatched :: forall (dev :: (DeviceType, Nat)) (batchSize :: Nat). (IsValidDevice dev, 1 <= batchSize) => QModel dev -> QEncoding dev '[batchSize] -> QTensor dev '[batchSize, 1]

forwardPolicy :: forall (dev :: (DeviceType, Nat)). (GeluDTypeIsValid dev QDType, RandDTypeIsValid dev QDType, BasicArithmeticDTypeIsValid dev QDType, SumDTypeIsValid dev QDType, MeanDTypeValidation dev QDType, StandardFloatingPointDTypeValidation dev QDType, KnownDevice dev) => QModel dev -> QEncoding dev ('[] :: [Nat]) -> QTensor dev '[1]

forwardPolicyBatched :: forall (dev :: (DeviceType, Nat)) (batchSize :: Nat). (OrdCond (CmpNat 1 batchSize) 'True 'True 'False ~ 'True, GeluDTypeIsValid dev QDType, RandDTypeIsValid dev QDType, BasicArithmeticDTypeIsValid dev QDType, SumDTypeIsValid dev QDType, MeanDTypeValidation dev QDType, StandardFloatingPointDTypeValidation dev QDType, KnownDevice dev) => QModel dev -> QEncoding dev '[batchSize] -> QTensor dev '[batchSize, 1]

forwardValue :: forall (dev :: (DeviceType, Nat)). IsValidDevice dev => QModel dev -> StateEncoding dev -> QTensor dev '[1]

fakeLoss :: forall (dev :: (DeviceType, Nat)) (ps :: [Type]). (IsValidDevice dev, ps ~ Parameters (QModel dev)) => QModel dev -> QTensor dev ('[] :: [Nat])

A loss for any model with 0 gradients everywhere. Can be used to ensure that all parameters have a gradient, if not all parameters are used in the real loss.

mkQModel :: forall (dev :: (DeviceType, Nat)). IsValidDevice dev => IO (QModel dev)

loadModel :: forall (dev :: (DeviceType, Nat)). IsValidDevice dev => FilePath -> IO (QModel dev)

saveModel :: forall (dev :: (DeviceType, Nat)). FilePath -> QModel dev -> IO ()

modelSize :: forall (dev :: (DeviceType, Nat)). QModel dev -> Int

runQ :: forall (dev :: (DeviceType, Nat)) s a. IsValidDevice dev => (s -> a -> QEncoding dev ('[] :: [Nat])) -> QModel dev -> s -> a -> QType

runQ' :: forall (dev :: (DeviceType, Nat)) s a. IsValidDevice dev => (s -> a -> QEncoding dev ('[] :: [Nat])) -> QModel dev -> s -> a -> QTensor dev '[1]

runBatchedPolicy :: forall (dev :: (DeviceType, Nat)) (batchSize :: Nat). (IsValidDevice dev, KnownNat batchSize) => QModel dev -> QEncoding dev '[batchSize] -> Tensor